WIP help menu
Adds a help menu bound to the '/' key to display a formatted list of
commands and their associated keybindings to aid in discovery.

To support this, a new COMMANDS constant is defined as static array of
instances of a newly defined `Command` trait. This trait exposes methods
to fetch the keybindings, description, detailed description, and the
implementation of the command.

Future work will try to unify the current command dispatching mechanism
(i.e. the big command switch) to reduce duplication.